+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 4.5.1 - 2026-01-26
 Ability to return any status codes and headers in `directReturn` if needed.
 
 ## 4.5.0 - 2026-01-25